+++ markdown.js
@@ -1,7 +1,7 @@
 import { Marked } from 'marked';
 import { markedHighlight } from 'marked-highlight';
 import hljs from 'highlight.js';
-import { text, sanitizeString } from '@shgysk8zer0/aegis';
+import { sanitizeString } from '@shgysk8zer0/aegis';
 
 export function createMDParser({
 	gfm = true,
@@ -25,8 +25,8 @@ export function createMDParser({
 		})
 	);
 
-	return (strings, ...args) => {
-		const parsed = marked.parse(text(strings, ...args), { gfm, breaks, silent });
+	return (...args) => {
+		const parsed = marked.parse(String.raw.apply(null, args), { gfm, breaks, silent });
 
 		return sanitizeString(parsed, {
 			allowElements, allowAttributes, allowCustomElements, allowUnknownMarkup,